feat(shokunin): new directory & minification files
