+++ src/utilities/mod.rs
@@ -1,0 +1,9 @@
+
+// Copyright © 2023 Shokunin Static Site Generator. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+/// The `directory` module contains functions for creating directories.
+pub mod directory;
+
+/// The `minification` module contains functions for minifying HTML files.
+pub mod minification;
