+++ tests/test_utilities.rs
@@ -4,19 +4,27 @@ mod tests {
     use comrak::ComrakOptions;
     use quick_xml::Writer;
     use regex::Regex;
-    use ssg::utilities::{
-        backup_file, cleanup_directory, create_comrak_options,
-        create_directory, directory, extract_front_matter,
-        find_html_files, format_header_with_id_class, minify_html,
-        minify_html_files, move_output_directory, to_title_case,
-        update_class_attributes, write_element, write_minified_html,
-    };
-    use std::io::Cursor;
+    use
+        ssg::utilities::{directory::{
+            backup_file,
+            cleanup_directory,
+            create_comrak_options,
+            create_directory,
+            directory,
+            extract_front_matter,
+            find_html_files,
+            format_header_with_id_class,
+            move_output_directory,
+            to_title_case,
+            update_class_attributes,
+            write_element
+        },
+        minification::{minify_html_files, minify_html,write_minified_html}};
     use std::{
         error::Error,
         fs::{self, File},
-        io::{self, Read, Write},
-        path::{Path, PathBuf},
+        io::{Cursor, self, Read, Write},
+        path::{Path, PathBuf}
     };
     use tempfile::tempdir;
